- moved WellKnownPaths from Nox.Abstractions to Nox.Solution - inverted depedencies so that Nox.Secrets depends on Nox.Solution, Secrets implementations depend on Nox.Scerets - Moved NoxSecretsResolver.cs to Nox.Lib - re-organized DI extensions
