+++ src/Nox.Generator.sln
@@ -85,8 +85,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nox.Secrets.Azure", "Nox.Se
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nox.Secrets.Hashicorp", "Nox.Secrets.Hashicorp\Nox.Secrets.Hashicorp.csproj", "{0B3974A4-121C-486B-964F-9037DAD3C5E8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nox.Secrets.Abstractions", "Nox.Secrets.Abstractions\Nox.Secrets.Abstractions.csproj", "{B77B8B5E-A0FD-45BB-9644-61878079149A}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -193,10 +191,6 @@ Global
 		{0B3974A4-121C-486B-964F-9037DAD3C5E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0B3974A4-121C-486B-964F-9037DAD3C5E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0B3974A4-121C-486B-964F-9037DAD3C5E8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B77B8B5E-A0FD-45BB-9644-61878079149A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B77B8B5E-A0FD-45BB-9644-61878079149A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B77B8B5E-A0FD-45BB-9644-61878079149A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B77B8B5E-A0FD-45BB-9644-61878079149A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -226,7 +220,6 @@ Global
 		{9C5DF8EC-0F0C-4FF0-86CC-3BE5181E6292} = {2B151195-53F6-45CD-84D5-F1E18CDFF1B0}
 		{D017DD5E-2DE1-462A-A891-CAB16C8BF940} = {2B151195-53F6-45CD-84D5-F1E18CDFF1B0}
 		{0B3974A4-121C-486B-964F-9037DAD3C5E8} = {2B151195-53F6-45CD-84D5-F1E18CDFF1B0}
-		{B77B8B5E-A0FD-45BB-9644-61878079149A} = {2B151195-53F6-45CD-84D5-F1E18CDFF1B0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0E72B357-9E1E-43B5-9868-8D2B74CB2AF0}
