+++ src/Nox.Lib/Extensions/ServiceCollectionExtension.cs
@@ -1,5 +1,6 @@
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 using Nox.Abstractions;
 using Nox.Secrets;
 using Nox.Secrets.Abstractions;
@@ -45,4 +46,17 @@ public static class ServiceCollectionExtension
             .Build();
     }
 
+    internal static IServiceCollection AddSecretsResolver(this IServiceCollection services)
+    {
+        services.AddPersistedSecretStore();
+        services.TryAddSingleton<INoxSecretsResolver, NoxSecretsResolver>();
+        return services;
+    }
+ 
+    internal static IServiceCollection AddPersistedSecretStore(this IServiceCollection services)
+    {
+        services.AddDataProtection();
+        services.AddSingleton<IPersistedSecretStore, PersistedSecretStore>();
+        return services;
+    }
 }
