+++ src/Nox.Lib/Nox.Lib.csproj
@@ -39,7 +39,8 @@
 		<ProjectReference Include="..\Nox.EntityFramework.Postgres\Nox.EntityFramework.Postgres.csproj" />
 		<ProjectReference Include="..\Nox.EntityFramework.Sqlite\Nox.EntityFramework.Sqlite.csproj" />
 		<ProjectReference Include="..\Nox.EntityFramework.SqlServer\Nox.EntityFramework.SqlServer.csproj" />
-		<ProjectReference Include="..\Nox.Secrets.Abstractions\Nox.Secrets.Abstractions.csproj" />
+		<ProjectReference Include="..\Nox.Secrets.Azure\Nox.Secrets.Azure.csproj" />
+		<ProjectReference Include="..\Nox.Secrets.Hashicorp\Nox.Secrets.Hashicorp.csproj" />
 		<ProjectReference Include="..\Nox.Secrets\Nox.Secrets.csproj" />
 		<ProjectReference Include="..\Nox.Solution\Nox.Solution.csproj" />
 		<ProjectReference Include="..\Nox.Abstractions\Nox.Abstractions.csproj" />
