+++ src/Nox.Secrets.Azure/AzureSecretsProvider.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 using Azure.Security.KeyVault.Secrets;
 using Nox.Secrets.Abstractions;
+using Nox.Secrets.Exceptions;
 
 namespace Nox.Secrets.Azure;
 
@@ -16,7 +17,7 @@ public class AzureSecretsProvider: ISecretsProvider
         }
         else
         {
-            throw new NoxSecretsException("VaultUrl is not a well formed Uri!");
+            throw new NoxSecretsException(ExceptionResources.MalformedVaultUri);
         }
     }
