+++ src/Nox.Secrets.Azure/Nox.Secrets.Azure.csproj
@@ -28,11 +28,26 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Nox.Abstractions\Nox.Abstractions.csproj" />
-      <ProjectReference Include="..\Nox.Secrets.Abstractions\Nox.Secrets.Abstractions.csproj" />
+      <ProjectReference Include="..\Nox.Secrets\Nox.Secrets.csproj" />
     </ItemGroup>
 
     <ItemGroup>
         <None Include="..\..\images\nox.png" Pack="true" PackagePath="" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <EmbeddedResource Update="ExceptionResources.resx">
+        <Generator>ResXFileCodeGenerator</Generator>
+        <LastGenOutput>ExceptionResources.Designer.cs</LastGenOutput>
+      </EmbeddedResource>
+    </ItemGroup>
+
+    <ItemGroup>
+      <Compile Update="ExceptionResources.Designer.cs">
+        <DesignTime>True</DesignTime>
+        <AutoGen>True</AutoGen>
+        <DependentUpon>ExceptionResources.resx</DependentUpon>
+      </Compile>
     </ItemGroup>
     
 </Project>
