+++ src/Nox.Secrets.Hashicorp/HashicorpSecretsResolver.cs
@@ -1,6 +1,7 @@
 using System.Net.Sockets;
 using Nox.Abstractions;
 using Nox.Secrets.Abstractions;
+using Nox.Secrets.Exceptions;
 using Nox.Solution;
 
 namespace Nox.Secrets.Hashicorp;
@@ -13,7 +14,7 @@ public class HashicorpSecretsResolver: ISecretsResolver
 
     public HashicorpSecretsResolver(IPersistedSecretStore store, SecretsServer secretsServer, string? storePrefix = null)
     {
-        if (secretsServer.Provider != SecretsServerProvider.HashicorpVault) throw new NoxSecretsException("Hashicorp secrets resolver can only be instantiated if provider is set to HashicorpVault");
+        if (secretsServer.Provider != SecretsServerProvider.HashicorpVault) throw new NoxSecretsException(ExceptionResources.InvalidProvider);
         _store = store;
         _storePrefix = "";
         if (!string.IsNullOrWhiteSpace(storePrefix)) _storePrefix = storePrefix + '.';
@@ -70,7 +71,8 @@ public class HashicorpSecretsResolver: ISecretsResolver
                     }
                     catch (Exception ex)
                     {
-                        throw new NoxSecretsException("Unable to retrieve secrets from the Hashicorp vault, are you connected to the internet and is the vault available?", ex);                        
+                        
+                        throw new NoxSecretsException(ExceptionResources.VaultUnavailable, ex);                        
                     }
 
                     break;
