+++ src/Nox.Secrets.Hashicorp/Nox.Secrets.Hashicorp.csproj
@@ -27,7 +27,22 @@
     <ItemGroup>
       <ProjectReference Include="..\Nox.Abstractions\Nox.Abstractions.csproj" />
       <PackageReference Include="VaultSharp" Version="1.13.0.1" />
-      <ProjectReference Include="..\Nox.Secrets.Abstractions\Nox.Secrets.Abstractions.csproj" />
+      <ProjectReference Include="..\Nox.Secrets\Nox.Secrets.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <EmbeddedResource Update="ExceptionResources.resx">
+        <Generator>ResXFileCodeGenerator</Generator>
+        <LastGenOutput>ExceptionResources.Designer.cs</LastGenOutput>
+      </EmbeddedResource>
+    </ItemGroup>
+
+    <ItemGroup>
+      <Compile Update="ExceptionResources.Designer.cs">
+        <DesignTime>True</DesignTime>
+        <AutoGen>True</AutoGen>
+        <DependentUpon>ExceptionResources.resx</DependentUpon>
+      </Compile>
     </ItemGroup>
     
 </Project>
