+++ src/Nox.Secrets/Abstractions/IPersistedSecretStore.cs
