+++ src/Nox.Secrets/Abstractions/ISecretsResolver.cs
