+++ src/Nox.Secrets/Nox.Secrets.csproj
@@ -43,9 +43,6 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\tests\Nox.Solution.Tests\Nox.Solution.Tests.csproj" />
-      <ProjectReference Include="..\Nox.Secrets.Abstractions\Nox.Secrets.Abstractions.csproj" />
-      <ProjectReference Include="..\Nox.Secrets.Azure\Nox.Secrets.Azure.csproj" />
-      <ProjectReference Include="..\Nox.Secrets.Hashicorp\Nox.Secrets.Hashicorp.csproj" />
+      <ProjectReference Include="..\Nox.Solution\Nox.Solution.csproj" />
     </ItemGroup>
 </Project>
