+++ src/Nox.Solution/Constants/CliWellKnownPaths.cs
@@ -1,6 +1,8 @@
-namespace Nox.Abstractions.Configuration;
+using System.IO;
 
-public class CliWellKnownPaths
+namespace Nox.Solution.Constants;
+
+public static class CliWellKnownPaths
 {
     public static readonly string CacheFile = Path.Combine(WellKnownPaths.CachePath, "NoxCliCache.json");
     public static readonly string WorkflowsCachePath = Path.Combine(WellKnownPaths.CachePath, "workflows");
