+++ tests/Nox.Secrets.Tests/CacheTests.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Nox.Abstractions;
-using Nox.Abstractions.Configuration;
 using Nox.Secrets.Abstractions;
+using Nox.Solution.Constants;
 using File = System.IO.File;
 
 namespace Nox.Secrets.Tests;
