+++ tests/Nox.Secrets.Tests/Nox.Secrets.Tests.csproj
@@ -28,9 +28,9 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\src\Nox.Lib\Nox.Lib.csproj" />
       <ProjectReference Include="..\..\src\Nox.Secrets.Azure\Nox.Secrets.Azure.csproj" />
       <ProjectReference Include="..\..\src\Nox.Secrets.Hashicorp\Nox.Secrets.Hashicorp.csproj" />
-      <ProjectReference Include="..\..\src\Nox.Secrets\Nox.Secrets.csproj" />
     </ItemGroup>
 
 </Project>
